refactor: resolve #758 - extract shared note-to-frequency logic from useComposerPlayback and MusicDSLParser
Closes #758
